improved login functions main.js to scrape url href and work with console parameters scrape_by_keywords.js to generate search requests and find new companies

+++ main.js
@@ -8,6 +8,7 @@ require('./login-xing.js');
 1. Company Name - name
 2. Xing Link - url
 3. Adresse (Straße) - street_address
+
 4. Adresse (PLZ) - post_index
 5. Adresse (Stadt) - city
 6. Adresse (Land) - country
@@ -33,7 +34,6 @@ var categoriesMap = {'over_10000': 9 };
 var syllables = ['BA','BE','BI','BO','BU','BY','CA','CE','CI','CO','CU','CY','DA','DE','DI','DO','DU','DY','FA','FE','FI','FO','FU','FY','GA','GE','GI','GO','GU','GY','HA','HE','HI','HO','HU','HY','JA','JE','JI','JO','JU','JY','KA','KE','KI','KO','KU','KY','LA','LE','LI','LO','LU','LY','MA','ME','MI', 'MO','MU','MY','NA','NE','NI','NO','NU','NY','PA','PE','PI','PO','PU','PY','QA','QE','QI','QO','QU','QY','RA','RE','RI','RO','RU','RY','SA','SE','SI','SO','SU','SY','TA','TE','TI','TO','TU','TY','VA','VE','VI','VO','VU','VY','WA','WE','WI','WO','WU','WY','XA','XE','XI','XO','XU','XY' ].reverse();
 var get_parameters=''; 
 
-
 process.argv.forEach(function (val, index, array) {
   console.log(index + ': ' + val);
   if (val.startsWith('country=')) {
@@ -41,7 +41,7 @@ process.argv.forEach(function (val, index, array) {
   }
 }); 
 
-countries.split(',').forEach(function (item, index) {
+/*countries.split(',').forEach(function (item, index) {
     //console.log(item, index, countriesMap[item]);
     get_parameters += '&filter.location[]='+countriesMap[item];
 });
@@ -49,8 +49,7 @@ countries.split(',').forEach(function (item, index) {
 
 var companies_req = 'https://www.xing.com/search/companies?filter.size[]=9' + get_parameters
 console.log('\ncompanies_req:', companies_req);
-
-process.exit();
+*/
 function getValidRequest(request, queue){ 
 	while ( request.url.includes('/employees') || 
 			request.url.includes('/industries')  ||
@@ -103,17 +102,28 @@ function addLinksToRequestQueue(links, requestQueue){
 	}
 	return requestQueue;
 }
-Apify.main(async () => {  
+Apify.main(async () => { 
+	const store = await Apify.openKeyValueStore('default');
+	const input = await store.getValue('INPUT-ge');
+	//console.log('\ncustom input:', input);
+	//process.exit(); 
 	// init variables from INPUT json file - apify_storage/key_value_stores/default/INPUT.json
-	const input = await Apify.getInput(); // https://sdk.apify.com/docs/api/apify#module_Apify.getInput
+	/*if (!custom_input) { 
+		const input = await Apify.getInput(); // https://sdk.apify.com/docs/api/apify#module_Apify.getInput
+	} else {
+		const input = custom_input;
+	}*/
+	//console.log('\n[custom] input:', input);
 	var concurrency =  parseInt(input.concurrency);
 	accounts = input.account; 
-	for (let item in accounts) {
+	/*for (let item in accounts) {
 	   console.log(item,':' ,accounts[item]);
-	}
+	}*/
 	let account = accounts[input.account_index];
 	console.log( '\n We select "account',  input.account_index,'" : ' ,account );
 	
+	//await login(account.username, account.password, input.cookieFile);
+	//console.log('after login...');
 	//process.exit();
 	var page_handle_max_wait_time = parseInt( input.page_handle_max_wait_time);
 	var max_requests_per_crawl =  parseInt( input.max_requests_per_crawl);
@@ -124,33 +134,62 @@ Apify.main(async () => {
 	var page_content=''; 
 	//console.log('\ninput.username:' , input.username, '\ninput.password:', input.password);
 	
-	const dataset = await Apify.openDataset('scraped-info');	
-    const requestQueue = await Apify.openRequestQueue();  
-	//requestQueue.addRequest({ url: 'https://www.xing.com/companies/daimlerag'});
-	//requestQueue.addRequest({ url: 'https://www.xing.com/companies/iav'});
+	// get countries and employees size from INPUT
+	if ( input.hasOwnProperty('crawl') ) {
+		input.crawl.country.split(',').forEach(function (item, index) {
+			console.log(index, 'country:', item); //, index, countriesMap[item]);
+			get_parameters += '&filter.location[]=' + countriesMap[item];
+		});
+		
+		input.crawl.empl_range.split(',').forEach(function (item, index) {
+			console.log(index, 'empl. range [category]:', input.crawl.empl_range);
+			get_parameters += '&filter.size[]=' + item; 
+		});
+	}
+	
+	//var companies_req = 'https://www.xing.com/search/companies?section=search&sc_o=companies_search_button' + get_parameters; //.substring(1);
+	//console.log('\ncompanies_req:', companies_req);
+	
+	const dataset = await Apify.openDataset('test_ds2');	
+    const requestQueue = await Apify.openRequestQueue('test_ds2');  
+	
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/4bag'});
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/iav'});
 	requestQueue.addRequest({ url: 'https://www.xing.com/companies/mercedes-amggmbh'});
-    
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/siselinternational'});
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/allianzdeutschlandag'});
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/mercedes-benzcustomersolutionsgmbh'});
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/intel'});
+	requestQueue.addRequest({ url: 'https://www.xing.com/companies/murataelektronikgmbh'});
+	
+	
+	//requestQueue.addRequest({ url: 'https://www.xing.com/search/companies?section=search'});
+    //requestQueue.addRequest({ url:  companies_req});
+	
 	//requestQueue.addRequest({ url: 'https://www.xing.com/companies' });
 	//requestQueue.addRequest({ url: 'https://www.xing.com/companies/recommendations' });
 	//requestQueue.addRequest({ url: 'https://www.xing.com/companies/recommendations?page=2'});
 	
 	//const pseudoUrls = [new Apify.PseudoUrl('https://www.xing.com/companies/[.+]')];
 	//const pseudoUrls = [new Apify.PseudoUrl(/https:\/\/www\.xing\.com\/companies\/(\w|-)+(?!\/jobs|\/report|\/affiliations|\/follower|\/search|\/reviews|\/industries|\/updates|\/employees)/gm)];
-	const pseudoUrls = [new Apify.PseudoUrl(/https:\/\/www\.xing\.com\/companies\/(\w|-)*/)];
+	//const pseudoUrls = [new Apify.PseudoUrl(/https:\/\/www\.xing\.com\/companies\/(\w|-)*/)];
 	// hint for multiple negative lookahead: https://stackoverflow.com/a/47281442/1230477
 	// link to test: https://regex101.com/r/JFIUff/1
-	
+//	console.log('Start logging-in...');
+	//await login(account.username, account.password, input.cookieFile, 1);
+	//console.log('Login is done before the crawler run to get fresh cookies written into file...');
+   
     const crawler = new Apify.PuppeteerCrawler({
         requestQueue, 
 		maxRequestsPerCrawl: max_requests_per_crawl,
         maxConcurrency: concurrency,
-		launchPuppeteerOptions: { slowMo: 45 } , 
+		launchPuppeteerOptions: { slowMo: 15 } , 
 		gotoFunction: async ({ request, page }) => { 			
 			try { 
 			    if (!login_flag) { // we login at the first request 
-					login_flag = true;  
-					//console.log('\n\n Request Queue:\n', requestQueue);
-					await login_page(page, account.username, account.password, input.cookieFile);					
+					//await set_cookie(page);
+					await login_page(page, account.username, account.password, input.cookieFile);
+					//console.log('Cookie is set for log-in!');					
 				} 	
 				await page.goto(request.url, { timeout: 60000 });
 			} catch (error){
@@ -158,18 +197,20 @@ Apify.main(async () => {
 			};  
 		},
         handlePageFunction: async ({ request, page }) => {
+			// we change the flag
+			if (!login_flag) { login_flag = true; }
+			
 			await page.waitFor(Math.ceil(Math.random() * page_handle_max_wait_time))
-			var company_name='';
+			var company_name='';		
 		
-		
-            try {
+            try { // get name of the company
 				var name_element = await page.$('h1.organization-name');
 				company_name = await (await name_element.getProperty('textContent')).jsonValue();
 				var company_name = company_name.trim();  
 			} catch(error){
 				//console.log(`\nNo company name in url: ${request.url}:\n`); //, error);
 			}
-			if (company_name) {
+			if (company_name) { // get info from page's html and saving into dataset
 				// section
 				var summary_text='';
 				try {
@@ -237,10 +278,13 @@ Apify.main(async () => {
 				} 
 				var website='';
 				try {
-					var website_element = await page.$('a[itemprop="url"]');
-					website = await (await website_element.getProperty('textContent')).jsonValue();
+					var website_element = await page.$('a[itemprop="url"]');  
+					website = await (await website_element.getProperty('href')).jsonValue(); 
+					if (!website || website == "Homepage") {
+						console.log(`problem, website: "${website}" \nfor url "${request.url}"`);
+					}	 
 				} catch(error){
-					//console.log(`\nFailure to get website  for url: ${request.url}: `, error);
+					//console.log(`\nFailure to get "website" for url: ${request.url}: `, error);
 				} 
 				console.log(`Company for ${request.url}: ${company_name}`);
 				var product_services = '';
@@ -297,47 +341,60 @@ Apify.main(async () => {
 					website: website,		
 				});
 			}
-			page_content = await page.content();
-			links_found  =  findAll(link_regex, page_content);
-			links_found_short =  findAll(short_link_regex, page_content);  			
 			
-			for (elem of links_found_short) {
-				links_found.add('https://www.xing.com'+elem);	 
-			} 
-			console.log('FOUND LINKS at "'+ request.url +'": (', links_found.size, ')\n', links_found );
 			
-			addLinksToRequestQueue(links_found, requestQueue); 
-			printRequestQueue(requestQueue);
-            //await Apify.utils.enqueueLinks({ page, selector: 'a', pseudoUrls, requestQueue });
+			// we gather links
+			if (request.url.includes('/search/companies')) {
+				page_content = await page.content();
+				links_found  =  findAll(link_regex, page_content);
+				links_found_short =  findAll(short_link_regex, page_content);  			
+				
+				for (elem of links_found_short) {
+					links_found.add('https://www.xing.com'+elem);	 
+				} 
+				console.log('FOUND LINKS at "'+ request.url +'": (', links_found.size, ')\n', links_found );
+				// add those links
+				addLinksToRequestQueue(links_found, requestQueue);
+				
+				// add a next companies request
+				let page = parseInt(request.url.split('page=')[1].split('&')[0]);	
+				if (isNaN(page)) { page=1; }	
+				page +=1;
+				console.log('\n next companies page (index) to crawl :', page);
+			}
+			
+			// print queue
+			var { totalRequestCount, handledRequestCount, pendingRequestCount } = await requestQueue.getInfo();
+			console.log('RequestQueue\n handled:', handledRequestCount);
+			console.log(' pending:', pendingRequestCount);
+			console.log(' total:'  , totalRequestCount);
         },
         handleFailedRequestFunction: async ({ request }) => {
             console.log(`Request ${request.url} failed too many times`);
             await Apify.pushData({
                 '#debug': Apify.utils.createRequestDebugInfo(request),
             });
-        },
-        
+        },        
     });
 
     await crawler.run();
 
-	printRequestQueue(requestQueue);
-	console.log('\nDeleting requestQueue');
-	await requestQueue.delete();
+	//printRequestQueue(requestQueue);
+	if (input.deleteQueue) {
+		console.log('\nDeleting requestQueue');
+		await requestQueue.delete();
+	}
 	// await queue.addRequest(new Apify.Request({ url: 'http://example.com/foo/bar'}, { forefront: true });
 
-	console.log('\n******** Results ********');
-	/*var obj = { companies: [] };	 
-	await dataset.forEach(async (item, index) => {		
-	    if (item.name) { 
-			obj.companies.push(item);
-			console.log(`${index}. "${item.name}" ${item.url}`); 
-			//if (item.summary ) { console.log(item.summary ); }
-			//if (item.contact ) { console.log(item.contact ); }
-		}
-	});*/
+	console.log('\n******** Results ********');	 
+	// print final queue
+	var { totalRequestCount, handledRequestCount, pendingRequestCount } = await requestQueue.getInfo();
+	console.log('Final RequestQueue\n handled:', handledRequestCount);
+	console.log(' pending:', pendingRequestCount);
+	console.log(' total:'  , totalRequestCount);
+	
 	const data = await dataset.getData().then((response) => response.items);
-	await Apify.setValue('OUTPUT', data);
+	await Apify.setValue(input.output, data);
 	 
 	const { itemCount } = await dataset.getInfo();
 	console.log('\nTotal items in dataset: ', itemCount);
